+++ expo-maps/android/src/main/java/expo/modules/maps/Gestures.kt
