+++ expo-maps/android/src/main/java/expo/modules/maps/Polylines.kt
@@ -1,6 +1,6 @@
 package expo.modules.maps
 
 interface Polylines {
-  fun setPolylines(polylinesObjects: Array<PolylineObject>)
+  fun setPolylines(polylineObjects: Array<PolylineObject>)
   fun detachAndDeletePolylines()
 }
