+++ expo-maps/ios/ExpoMaps/AppleMapsGestures.swift
@@ -1,4 +1,5 @@
 import MapKit
+
 final class AppleMapsGestures : Gestures {
   
   internal var mapView: MKMapView;
