+++ expo-maps/ios/ExpoMaps/ExpoGoogleMapsModule.swift
@@ -1,6 +1,7 @@
 import ExpoModulesCore
 
 public class ExpoGoogleMapsModule: Module {
+  
   public func definition() -> ModuleDefinition {
     name("ExpoGoogleMaps")
 
@@ -8,21 +9,27 @@ public class ExpoGoogleMapsModule: Module {
       view {
         GoogleMapsView()
       }
+      
       prop("enableRotateGestures") { (view: GoogleMapsView, enable: Bool) in
-          view.setEnabledRotateGestures(enabled: enable)
+        view.setEnabledRotateGestures(enabled: enable)
       }
+      
       prop("enableScrollGestures") { (view: GoogleMapsView, enable: Bool) in
-          view.setEnabledScrollGestures(enabled: enable)
+        view.setEnabledScrollGestures(enabled: enable)
       }
+      
       prop("enableTiltGestures") { (view: GoogleMapsView, enable: Bool) in
-          view.setEnabledTiltGestures(enabled: enable)
+        view.setEnabledTiltGestures(enabled: enable)
       }
+      
       prop("enableZoomGestures") { (view: GoogleMapsView, enable: Bool) in
-          view.setEnabledZoomGestures(enabled: enable)
+        view.setEnabledZoomGestures(enabled: enable)
       }
+      
       prop("mapType") { (view: GoogleMapsView, mapType: MapType) in
         view.setMapType(mapType: mapType)
       }
+      
       prop("jsonStyleString") { (view: GoogleMapsView, jsonStyleString: String) in
         view.setMapStyle(jsonStyleString: jsonStyleString)
       }
