+++ expo-maps/ios/ExpoMaps/ExpoMapView.swift
@@ -1,13 +1,9 @@
 protocol ExpoMapView: UIView {
   init()
-  
   init?(coder: NSCoder)
   
   func setMapType(mapType: MapType)
-  
   func setMarkers(markerObjects: [MarkerObject])
-  
   func setPolygons(polygonObjects: [PolygonObject])
-  
   func setPolylines(polylineObjects: [PolylineObject])
 }
