+++ expo-maps/ios/ExpoMaps/GoogleMapsView.swift
@@ -1,10 +1,10 @@
 import GoogleMaps
 
 public final class GoogleMapsView: UIView, ExpoMapView {
+  
   private let mapView: GMSMapView
   private let markers: GoogleMapsMarkers
   private let gestures: GoogleMapsGestures
-  private let markers: GoogleMapsMarkers
   private let polygons: GoogleMapsPolygons
   private let polylines: GoogleMapsPolylines
