+++ expo-maps/ios/ExpoMaps/Markers.swift
@@ -1,6 +1,5 @@
 protocol Markers {
   func detachAndDeleteMarkers()
-  
   func setMarkers(markerObjects: [MarkerObject])
 }
