+++ expo-maps/ios/ExpoMaps/Polygons.swift
@@ -1,5 +1,4 @@
 protocol Polygons {
   func detachAndDeletePolygons()
-  
   func setPolygons(polygonObjects: [PolygonObject])
 }
