+++ expo-maps/ios/ExpoMaps/Polylines.swift
@@ -1,5 +1,4 @@
 protocol Polylines {
   func detachAndDeletePolylines()
-  
   func setPolylines(polylineObjects: [PolylineObject])
 }
